Finished Move Detection
The board class can now detect all legal moves in a position.

Begun the Bot class to make moves.

The next step will be to have the bot track moves in algebraic notation to allow for moves like en passant and promotion to be detected.

After this, the bot will then need to be able to evaluate a Board object to be able to make decisions of which moves to make.